BUG: #33 Comment - getCommentById 해결 - comment 변수 타입 Long -> Integer 변경

+++ springboot-app/src/main/java/com/uplus/eureka/comment/model/dto/Comment.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Comment {
-    private Long commentId;  // comment_id
+    private Integer commentId;  // comment_id
     private Integer userId;  // user_id
     private String randomNickname; // random_nickname
     private String commentText; // comment_text
